Move CustomerGuestType to CoreBundle
