+++ src/Sylius/Bundle/CoreBundle/spec/Form/EventSubscriber/SetCustomerFormSubscriberSpec.php
@@ -9,12 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\UserBundle\Form\EventSubscriber;
+namespace spec\Sylius\Bundle\CoreBundle\Form\EventSubscriber;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
-use Sylius\Component\Customer\Model\CustomerInterface;
+use Sylius\Component\Core\Model\CustomerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
@@ -32,7 +32,7 @@ final class SetCustomerFormSubscriberSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\UserBundle\Form\EventSubscriber\SetCustomerFormSubscriber');
+        $this->shouldHaveType('Sylius\Bundle\CoreBundle\Form\EventSubscriber\SetCustomerFormSubscriber');
     }
 
     function it_is_an_event_subscriber()
