+++ src/Sylius/Bundle/UserBundle/Resources/config/doctrine/model/Customer.orm.xml
@@ -17,7 +17,7 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="Sylius\Component\User\Model\Customer" table="sylius_customer">
+    <mapped-superclass name="Sylius\Component\Customer\Model\Customer" table="sylius_customer">
         <id name="id" column="id" type="integer">
             <generator strategy="AUTO" />
         </id>
@@ -41,7 +41,7 @@
             <gedmo:timestampable on="update"/>
         </field>
 
-        <many-to-many field="groups" target-entity="Sylius\Component\User\Model\GroupInterface">
+        <many-to-many field="groups" target-entity="Sylius\Component\Customer\Model\GroupInterface">
             <cascade>
                 <cascade-all />
             </cascade>
