+++ src/Sylius/Bundle/UserBundle/Resources/config/services.xml
@@ -200,9 +200,6 @@
 
         <!-- Form event subscribers-->
         <service id="sylius.form.event_subscriber.add_user_type" class="%sylius.form.event_subscriber.add_user_type.class%" />
-        <service id="sylius.form.event_subscriber.set_customer" class="%sylius.form.event_subscriber.set_customer_type.class%">
-            <argument type="service" id="sylius.repository.customer" />
-        </service>
 
         <!-- Validators -->
         <service id="validator.unique.registered_user" class="%sylius.user.validator.unique_registered_user.class%">
